colexec: move the contents of two files into the templates
Two structs were defined in separate files. Now their declaration has
been moved into the corresponding template files.

Release note: None
