+++ pkg/sql/colexec/selection_ops_tmpl.go
@@ -170,6 +170,22 @@ func _SEL_LOOP(_HAS_NULLS bool) { // */}}
 	// {{/*
 } // */}}
 
+// selConstOpBase contains all of the fields for binary selections with a
+// constant, except for the constant itself.
+type selConstOpBase struct {
+	OneInputNode
+	colIdx         int
+	decimalScratch decimalOverloadScratch
+}
+
+// selOpBase contains all of the fields for non-constant binary selections.
+type selOpBase struct {
+	OneInputNode
+	col1Idx        int
+	col2Idx        int
+	decimalScratch decimalOverloadScratch
+}
+
 // {{define "selConstOp"}}
 type _OP_CONST_NAME struct {
 	selConstOpBase
